+++ .github/ci-coverage-allowlist.yml
@@ -49,18 +49,15 @@ test_paths:
     paths:
       - tests/code_coverage_tests/test_aio_http_image_conversion.py
   - reason: >-
-      What is left of a second mirror that sat beside tests/test_litellm and ran nowhere. Its
-      other 30 files moved into the real mirror on 2026-08-20 and now run; these four cannot,
-      because each shares a filename with a live test whose contents are disjoint from it, so
-      landing them means merging test bodies rather than moving a file. Measured on the same
-      date: test_common_utils.py holds 15 tests the live file does not, test_oci_chat_transformation
-      13, test_deepseek_chat_transformation 12, and test_discoverable_endpoints 5. Revisit by
-      merging each into its twin, which is a content review, not a move
+      The last file of a second mirror that sat beside tests/test_litellm and ran nowhere. Its
+      other 33 files landed in the real mirror during August 2026, 30 as moves and 3 by merging
+      their bodies into the live file of the same name. This one cannot follow either route yet:
+      its live twin was rewritten from 1268 lines to 9434, and of the 19 tests here 5 have no
+      counterpart while 25 assertions fail against today's code, so what survives that rewrite
+      is a judgement about the endpoints, not a merge. Revisit by deciding which of the five
+      behaviours still hold
     paths:
-      - tests/litellm/llms/deepseek/chat/test_deepseek_chat_transformation.py
-      - tests/litellm/llms/oci/chat/test_oci_chat_transformation.py
       - tests/litellm/proxy/_experimental/mcp_server/test_discoverable_endpoints.py
-      - tests/litellm/proxy/management_endpoints/test_common_utils.py
   - reason: >-
       No job invokes this suite and its files mix pure transformation tests with ones driving live
       vendor vector stores, so assigning them needs a per-file decision
